requirements.txt
langchain
langchain_community
langchain_openai
langchain_core
langchain-chroma
chainlit
python-dotenv
pypdf
typesense
faiss-cpu